requirements.txt
bearlibterminal>=0.15.2
Pillow>=4.2.1
